trying to fix make errors

+++ include/linux/syscalls.h
@@ -846,5 +846,5 @@ asmlinkage long sys_open_by_handle_at(int mountdirfd,
 asmlinkage long sys_setns(int fd, int nstype);
 
 /* rtes sys_xxxx below this line */
-asmlinkage int sys_calc(const char* param1, const char* param2, char operation, char* result);
+asmlinkage long sys_calc(const char* param1, const char* param2, char operation, char* result);
 #endif
